Create Credit_Card_Fraud.py
Look what I have for you! Another complete project in Machine Learning! We will be building a Credit Card Fraud Detection System from scratch! It is going to be a very interesting project to learn!
For this project, we will be using the several methods of Anomaly detection with Probability Densities. 

We will be implementing the two major algorithms namely,
1. A local out wire factor to calculate anomaly scores.
2. Isolation forced algorithm. 

To get started we will first build a dataset of over 280,000 credit card transactions to work on!